drop custom prefix option
