+++ .github/workflows/emscripten.yml
@@ -7,14 +7,13 @@ jobs:
     defaults:
       run:
         shell: bash
-    name: Emscripten-C++${{matrix.std}}-${{matrix.build_type}}-${{matrix.lib}}-${{matrix.extra}}
+    name: Emscripten-C++${{matrix.std}}-${{matrix.build_type}}-${{matrix.lib}}
     runs-on: ubuntu-latest
     container: emscripten/emsdk
     strategy:
       fail-fast: true
       matrix:
         build_type: [Release, Debug]
-        extra: [no-custom-prefix, custom-prefix]
         lib: [static]
         std: [98, 11, 14, 17, 20]
 
@@ -49,7 +48,6 @@ jobs:
             -DCMAKE_FIND_ROOT_PATH_MODE_PROGRAM=NEVER \
             -DCMAKE_INSTALL_PREFIX=${{github.workspace}}/install \
             -DCMAKE_RANLIB=$(which emranlib) \
-            -DWITH_CUSTOM_PREFIX=${{matrix.extra == 'custom-prefix'}} \
             -G Ninja \
             -Werror
