+++ CMakeLists.txt
@@ -37,7 +37,6 @@ include (GNUInstallDirs)
 option (BUILD_SHARED_LIBS "Build shared libraries" ON)
 option (PRINT_UNSYMBOLIZED_STACK_TRACES
   "Print file offsets in traces instead of symbolizing" OFF)
-option (WITH_CUSTOM_PREFIX "Enable support for user-generated message prefixes" ON)
 option (WITH_GFLAGS "Use gflags" ON)
 option (WITH_GTEST "Use Google Test" ON)
 option (WITH_PKGCONFIG "Enable pkg-config support" ON)
@@ -694,10 +693,6 @@ if (CYGWIN OR WIN32)
   target_compile_definitions (glog PUBLIC GLOG_NO_ABBREVIATED_SEVERITIES)
 endif (CYGWIN OR WIN32)
 
-if (WITH_CUSTOM_PREFIX)
-  target_compile_definitions (glog PUBLIC GLOG_CUSTOM_PREFIX_SUPPORT)
-endif (WITH_CUSTOM_PREFIX)
-
 set_target_properties (glog PROPERTIES PUBLIC_HEADER "${GLOG_PUBLIC_H}")
 
 target_include_directories (glog BEFORE PUBLIC
@@ -780,27 +775,6 @@ if (BUILD_TESTING)
   )
 
   target_link_libraries (logging_unittest PRIVATE ${_GLOG_TEST_LIBS})
-
-  if (WITH_CUSTOM_PREFIX)
-    add_executable (logging_custom_prefix_unittest
-      src/logging_custom_prefix_unittest.cc
-    )
-
-    target_link_libraries (logging_custom_prefix_unittest PRIVATE ${_GLOG_TEST_LIBS})
-
-    add_test (NAME logging_custom_prefix
-              COMMAND logging_custom_prefix_unittest)
-
-    # FIXME: Skip flaky test
-    set_tests_properties (logging_custom_prefix PROPERTIES SKIP_REGULAR_EXPRESSION
-      "Check failed: time_ns within LogTimes::LOG_PERIOD_TOL_NS of LogTimes::LOG_PERIOD_NS")
-
-    if (APPLE)
-      # FIXME: Skip flaky test
-      set_property (TEST logging_custom_prefix APPEND PROPERTY SKIP_REGULAR_EXPRESSION
-        "unexpected new.*PASS\nTest with golden file failed. We'll try to show the diff:")
-    endif (APPLE)
-  endif (WITH_CUSTOM_PREFIX)
 
   add_executable (stl_logging_unittest
     src/stl_logging_unittest.cc
